Fixed the spelling errors
I Fixed lines 78 and 103
I'll download the file tomorrow and test it. I know the .format() is the in the string module, it's default onto any python file so there's no need to import anything. Maybe it can't to the variable, it might have to be a fixed string within the ()'s