The .format() was bugged
I could not find a way to get the .format(server.name) to work, I installed all the modules that the other cogs had and it was working fine for them:
welcome.py
economy.py

Kept getting an Attribute error, so I want to say it involved how we nested and defined our variables, I really don't know. However, I used a more ugly alternative, but it has been tested and worked on! Sorry for the delay and hope this fixes up those bugs!
Also added the checks cog, if you find no use for it then you can remove it.